change default units to check every N mins
Thanks for creating this plugin.

In order to keep the site fast for everyone we're now rate limiting requests to /studies. We advise that the page should be checked no more than once per min per user. As a result, the check interval should be in mins rather than seconds.